+++ DemoSnippets/ItemToAdd.cs
@@ -6,11 +6,11 @@
         {
             this.Label = string.Empty;
             this.Snippet = string.Empty;
-            this.TabLabel = string.Empty;
+            this.Tab = string.Empty;
         }
 
         public string Label { get; set; }
         public string Snippet { get; set; }
-        public string TabLabel { get; set; }
+        public string Tab { get; set; }
     }
 }