+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Road map
 
 - [x] Support installation on VS2019
-- [ ] Allow setting which tab in the toolbox snippets are added to
+- [x] Allow setting which tab in the toolbox snippets are added to
 - [ ] Allow processing multiple .demosnipptes files at once
 - [ ] Open to suggestions...